Amélioration de la recherche de diff Ajout de log Ajout des progress bars en html
