+++ src/AppBundle/Business/Exporter/AbstractExporter.php
@@ -27,20 +27,24 @@ abstract class AbstractExporter
         'Multimedia'
     ];
 
+    protected $userGroup;
     public $arrayEntityExport = [];
 
     /**
      * @param array  $datas
      * @param string $exportPath
+     * @param string $userGroup
      */
-    public function __construct($datas, $exportPath)
+    public function __construct($datas, $exportPath, $userGroup)
     {
         $this->datas = $datas;
         $this->exportPath = $exportPath;
-        if (!is_dir($this->exportPath)) {
+        $this->userGroup = $userGroup;
+        UtilityService::createDir($exportPath, $userGroup);
+        /*if (!is_dir($this->exportPath)) {
             mkdir($this->exportPath, 02774);
             chgrp($this->exportPath, UtilityService::getFileGroup($this->exportPath));
-        }
+        }*/
         foreach ($this->entitiesName as $className) {
             $entityExporterConstructor = '\\AppBundle\\Business\\Exporter\\Entity\\'.ucfirst($className).'Exporter';
             /* @var $entityExporter \AppBundle\Business\Exporter\Entity\AbstractEntityExporter */
