+++ src/AppBundle/Controller/BackendController.php
@@ -253,7 +253,7 @@ class BackendController extends Controller
         $diffHandler = $exportManager->getDiffHandler();
         $diffs = $diffHandler->getDiffsFile();
         $diffs->deleteChoices();
-        $exportManager->launchDiffProcess();
+        //$exportManager->launchDiffProcess();
         $this->get('session')->clear();
         $response = new JsonResponse();
         $response->setData(['deleteDiffs' => true]);
