+++ src/AppBundle/Manager/ExportManager.php
@@ -33,8 +33,6 @@ class ExportManager
      */
     private $collectionCode = null;
 
-    /** @var \AppBundle\Manager\DiffManager */
-    protected $diffManager;
     /**
      * @var integer
      */
@@ -76,7 +74,6 @@ class ExportManager
      * @param ManagerRegistry      $managerRegistry
      * @param Session              $sessionManager
      * @param GenericEntityManager $genericEntityManager
-     * @param DiffManager          $diffManager
      * @param int                  $maxItemPerPage
      * @param DiffComputer         $diffComputer
      * @param string               $userGroup
@@ -85,7 +82,6 @@ class ExportManager
         ManagerRegistry $managerRegistry,
         Session $sessionManager,
         GenericEntityManager $genericEntityManager,
-        DiffManager $diffManager,
         $maxItemPerPage,
         DiffComputer $diffComputer,
         $userGroup
@@ -93,7 +89,6 @@ class ExportManager
         $this->managerRegistry = $managerRegistry;
         $this->sessionManager = $sessionManager;
         $this->genericEntityManager = $genericEntityManager;
-        $this->diffManager = $diffManager;
         $this->maxItemPerPage = $maxItemPerPage;
         $this->diffComputer = $diffComputer;
         $this->userGroup = $userGroup;
@@ -124,7 +119,6 @@ class ExportManager
         if (is_null($this->collection)) {
             throw new \Exception('Can\'t found the collection with collectionCode = '.$this->collectionCode);
         } else {
-            $this->diffManager->init($this->collection);
             $this->diffHandler = new DiffHandler($this->user->getDataDirPath(), $this->collection, $this->userGroup);
 
             if (!$this->getDiffHandler()->shouldSearchDiffs()) {
@@ -138,26 +132,6 @@ class ExportManager
         }
 
         return $this;
-    }
-
-    /**
-     * @return array
-     */
-    public function launchDiffProcess()
-    {
-        if ($this->getDiffHandler()->shouldSearchDiffs()) {
-            $diffs = $this->diffManager->searchDiffs();
-            $diffComputer = $this->diffComputer->init($this->collection, $diffs);
-            $data = $diffComputer->getAllDatas();
-            $this->getDiffHandler()->saveDiffs($data);
-            $this->getDiffHandler()->getDiffsFile()->searchDiffs = false;
-        } else {
-            $data = $this->getDiffHandler()->getDiffsFile()->getData();
-        }
-
-        $data['selectedSpecimens'] = $this->selectedSpecimensHandler->getData();
-
-        return $data;
     }
 
     /**
@@ -466,10 +440,10 @@ class ExportManager
         $datasWithChoices = $this->prepareExport($exportPrefs);
         switch ($type) {
             case 'dwc':
-                $exporter = new DwcExporter($datasWithChoices, $this->getExportDirPath());
+                $exporter = new DwcExporter($datasWithChoices, $this->getExportDirPath(), $this->userGroup);
                 break;
             case 'csv':
-                $exporter = new CsvExporter($datasWithChoices, $this->getExportDirPath());
+                $exporter = new CsvExporter($datasWithChoices, $this->getExportDirPath(), $this->userGroup);
         }
         if ($exporter instanceof AbstractExporter) {
             return $exporter->generate($this->user->getPrefs());
@@ -487,8 +461,7 @@ class ExportManager
         $this->exportPrefs = $exportPrefs;
         $catalogNumbers = $this->sessionManager->get('catalogNumbers');
         $datas = $this->genericEntityManager->getEntitiesLinkedToSpecimens($this->exportPrefs->getSideForChoicesNotSet(),
-            $this->collection,
-            $catalogNumbers);
+            $this->collection, $catalogNumbers);
         $datasWithChoices = $this->getArrayDatasWithChoices($datas);
 
         return $datasWithChoices;
