+++ src/AppBundle/Resources/config/services.yml
@@ -4,15 +4,15 @@ services:
         arguments:
             managerRegistry: "@doctrine"
             maxNbSpecimenPerPass: "%max_nb_specimen_par_pass%"
+#    diff.manager:
+#        class: AppBundle\Manager\OldDiffManager
+#        arguments:
+#            managerRegistry: "@doctrine"
+#            recolnat_alias: "%database_user%"
+#            recolnat_diff_alias: "%database_user_diff%"
+
     diff.manager:
         class: AppBundle\Manager\DiffManager
-        arguments:
-            managerRegistry: "@doctrine"
-            recolnat_alias: "%database_user%"
-            recolnat_diff_alias: "%database_user_diff%"
-
-    diff.newmanager:
-        class: AppBundle\Manager\NewDiffManager
         arguments:
             managerRegistry: "@doctrine"
             recolnatAlias: "%database_user%"
@@ -29,7 +29,6 @@ services:
             managerRegistry: "@doctrine"
             sessionManager: "@session"
             genericEntityManager: "@genericentitymanager"
-            diffManager: "@diff.manager"
             maxItemPerPage: "%knp_paginator.page_range%"
             diffComputer: "@diff.computer"
             userGroup: "%user_group%"
