--- src/AppBundle/Resources/views/Front/generateDiff.html.twig
@@ -1,1 +1,0 @@
-{% extends '@App/base.html.twig' %}
