+++ src/AppBundle/Resources/views/Front/index.html.twig
@@ -25,50 +25,4 @@
             {% endfor %}
         {% endif %}
     </div>
-    {% include "@App/Front/partial/searchDiffModal.html.twig" %}
 {% endblock %}
-{#
-{% block javascripts %}
-    <script type="text/javascript">
-        $(document).ready(function () {
-            $("a.searchDiffButton").on('click', function (event) {
-                event.preventDefault();
-                $("#modal").modal({
-                    backdrop: 'static',
-                    keyboard: false
-                }).modal('show');
-
-                var progressbar = $("#modal").find(".progress-bar");
-                var steps;
-
-                try {
-                    var source = new EventSource($(this).attr('href'));
-                    source.onopen = function () {
-                        return;
-                    };
-                    source.addEventListener('steps', function (evt) {
-                        console.log(evt);
-                        steps = parseInt(evt.data);
-                    });
-                    source.addEventListener('step', function (evt) {
-                        data = jQuery.parseJSON(evt.data);
-                        progress(parseInt(data.count) / steps * 100, progressbar, data.step);
-                        if (parseInt(data.count) === steps) {
-                            source.close();
-                            window.location.reload();
-                        }
-                    });
-
-                } catch (e) {
-                    console.log(e);
-                }
-            });
-            function progress(percent, $element, label) {
-                $element.css('width', percent + "%").html(Math.ceil(percent) + "%");
-                $element.parent().next().html(Translator.trans(label));
-            }
-        });
-
-
-    </script>
-{% endblock %}#}
