+++ src/AppBundle/Resources/views/Front/partial/searchDiffModal.html.twig
@@ -2,21 +2,24 @@
     <div class="modal-dialog">
         <div class="modal-content">
             <div class="modal-header">
-                <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+                <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span
+                            aria-hidden="true">&times;</span></button>
                 <h4 class="modal-title">{{ 'titre.recherche.diffs' | trans }}</h4>
             </div>
             <div class="modal-body">
-                <div class="progress">
-                    <div class="progress-bar progress-bar-info progress-bar-striped" role="progressbar" aria-valuenow="0" aria-valuemin="0" aria-valuemax="100">
+                <div class="js-progress-template" style="display: none">
+                    <div class="progress-step-title js-progress-step-title"></div>
+                    <div class="progress">
+                        <div class="progress-bar progress-bar-info progress-bar-striped" role="progressbar"
+                             aria-valuenow="0" aria-valuemin="0" aria-valuemax="100">
+                        </div>
 
                     </div>
-
                 </div>
-                <div class="etape"></div>
             </div>
             <!--<div class="modal-footer">
-                <button type="button" class="btn btn-default" data-dismiss="modal">{{'modal.cancel' | trans}}</button>
-                <button type="button" class="btn btn-primary confirm">{{'modal.agree' | trans}}</button>
+                <button type="button" class="btn btn-default" data-dismiss="modal">{{ 'modal.cancel' | trans }}</button>
+                <button type="button" class="btn btn-primary confirm">{{ 'modal.agree' | trans }}</button>
             </div>-->
         </div><!-- /.modal-content -->
     </div><!-- /.modal-dialog -->
