+++ src/AppBundle/Resources/views/Front/searchDiffError.html.twig
@@ -1,0 +1,12 @@
+{% extends '@App/base.html.twig' %}
+
+{% block body %}
+    <div class="container-fluid">
+        <div class="bs-callout bs-callout-danger">
+            <h4>{{ "error.occured" | trans({}, 'exceptions') }}</h4>
+            {{ 'searchDiffError' | trans({}, 'exceptions') }}
+        </div>
+    </div>
+
+
+{% endblock %}
