+++ src/AppBundle/Resources/views/Front/viewFile.html.twig
@@ -145,10 +145,6 @@
                                         {{ 'delete.diffs'|trans }}
                                     </a>
                                 </p>
-                                <p>
-                                    <a href="{{ path('generateDiff', {'collectionCode' : collectionCode, 'compt' : 5}) }}"
-                                       class="btn btn-primary" role="button">Generate diff</a>
-                                </p>
                                 {% endif %}
                             </td>
                         {% endif %}
