+++ src/AppBundle/Tests/Units/Manager/ExportManager.php
@@ -50,7 +50,6 @@ class ExportManager extends Units\Test
             $managerRegistry,
             $this->container->get('session'),
             $this->container->get('genericentitymanager'),
-            $this->container->get('diff.manager'),
             $this->container->getParameter('maxitemperpage')[1],
             $this->container->get('diff.computer'),
             $this->container->getParameter('user_group')
